Split several functions out to new files for better organisation
