+++ config.m4
@@ -82,7 +82,7 @@ if test "$PHP_SCOUTAPM" != "no"; then
 
   PHP_SCOUTAPM_CFLAGS="-DZEND_ENABLE_STATIC_TSRMLS_CACHE=1 $STD_CFLAGS $MAINTAINER_CFLAGS"
 
-  PHP_NEW_EXTENSION(scoutapm, zend_scoutapm.c scout_curl_wrapper.c scout_file_wrapper.c scout_pdo_wrapper.c,
+  PHP_NEW_EXTENSION(scoutapm, zend_scoutapm.c scout_functions.c scout_recording.c scout_utils.c scout_curl_wrapper.c scout_file_wrapper.c scout_pdo_wrapper.c,
         $ext_shared,,$PHP_SCOUTAPM_CFLAGS,,yes)
 fi
