+++ package.xml
@@ -45,6 +45,9 @@
         <dir name="/">
             <file name="config.m4" role="src" />
             <file name="zend_scoutapm.c" role="src" />
+            <file name="scout_utils.c" role="src" />
+            <file name="scout_functions.c" role="src" />
+            <file name="scout_recording.c" role="src" />
             <file name="scout_curl_wrapper.c" role="src" />
             <file name="scout_file_wrapper.c" role="src" />
             <file name="scout_pdo_wrapper.c" role="src" />
