+++ scout_extern.h
@@ -17,9 +17,11 @@ extern int handler_index_for_function(const char *function_to_lookup);
 extern const char* determine_function_name(zend_execute_data *execute_data);
 extern const char *unique_resource_id(const char *scout_wrapper_type, zval *resource_id);
 extern const char *unique_class_instance_id(zval *class_instance);
+extern void safely_copy_argument_zval_as_scalar(zval *original_to_copy, zval *destination);
 
 ZEND_EXTERN_MODULE_GLOBALS(scoutapm);
 extern indexed_handler_lookup handler_lookup[];
+extern const int handler_lookup_size;
 extern zif_handler original_handlers[];
 
 #endif /* SCOUT_API_H */
